garage door entities have changed
The MyQ API is getting locked down and the old v4 API is gone. This
integrations with the MyQ platform to fail. MyQ requires a 3rd part to
pay to have access to their API officially but that doesn't make sense
for a self-hosted system like what is used for home automation.